app/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytest-django==4.1.0
pytest==6.2.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
pandas==0.25.0
protobuf==3.9.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
# tensorboard==1.14.0
tensorflow==1.13.1
# tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
whitenoise==4.1.3
wrapt==1.11.2
xlrd==1.2.0
django-crispy-forms==1.10.0